refactor: extract fitness heatmap client module
